moved code from hievpy.py into __init__.py to simplify import
